testbot/requirements.txt
python-telegram-bot==20.0
nest_asyncio==1.5.4
